Refactored run_download factory function from download.py to daac_data_subscriber.py to fix circular dependency workaround
